+++ SKPhotoBrowser/SKButtons.swift
@@ -9,40 +9,36 @@
 import Foundation
 
 // helpers which often used
-private let bundle = NSBundle(forClass: SKPhotoBrowser.self)
+private let bundle = Bundle(for: SKPhotoBrowser.self)
 
 class SKButton: UIButton {
     var showFrame: CGRect!
     var hideFrame: CGRect!
     var insets: UIEdgeInsets {
-
-
-        return UI_USER_INTERFACE_IDIOM() == .Phone
-            ?  UIEdgeInsets(top: 15.25, left: 15.25, bottom: 15.25, right: 15.25) : UIEdgeInsets(top: 12, left: 12, bottom: 12, right: 12)
+        if UI_USER_INTERFACE_IDIOM() == .phone {
+            return UIEdgeInsets(top: 15.25, left: 15.25, bottom: 15.25, right: 15.25)
+        } else {
+            return UIEdgeInsets(top: 12, left: 12, bottom: 12, right: 12)
+        }
     }
     var size: CGSize = CGSize(width: 44, height: 44)
     var margin: CGFloat = 5
-    
     var buttonTopOffset: CGFloat { return 5 }
     
-    func setup(imageName: String) {
-        backgroundColor = .clearColor()
-        tintColor = SKPhotoBrowserOptions.textAndIconColor
+    func setup(_ imageName: String) {
+        backgroundColor = .clear
         imageEdgeInsets = insets
-//        clipsToBounds = true
         translatesAutoresizingMaskIntoConstraints = true
-        autoresizingMask = [.FlexibleBottomMargin, .FlexibleLeftMargin, .FlexibleRightMargin, .FlexibleTopMargin]
+        autoresizingMask = [.flexibleBottomMargin, .flexibleLeftMargin, .flexibleRightMargin, .flexibleTopMargin]
         
         let image = UIImage(named: "SKPhotoBrowser.bundle/images/\(imageName)",
-                            inBundle: bundle, compatibleWithTraitCollection: nil)?.imageWithRenderingMode(.AlwaysTemplate) ?? UIImage()
-        setImage(image, forState: .Normal)
+                            in: bundle, compatibleWith: nil) ?? UIImage()
+        setImage(image, for: UIControlState())
     }
   
-    func updateFrame() { }
-  
-    func setFrameSize(size: CGSize) {
+    func setFrameSize(_ size: CGSize) {
         let newRect = CGRect(x: margin, y: buttonTopOffset, width: size.width, height: size.height)
-        self.frame = newRect
+        frame = newRect
         showFrame = newRect
         hideFrame = CGRect(x: margin, y: -20, width: size.width, height: size.height)
     }
@@ -60,9 +56,6 @@ class SKCloseButton: SKButton {
         showFrame = CGRect(x: margin, y: buttonTopOffset, width: size.width, height: size.height)
         hideFrame = CGRect(x: margin, y: -20, width: size.width, height: size.height)
     }
-    
-    override func updateFrame() {
-    }
 }
 
 class SKDeleteButton: SKButton {
@@ -77,11 +70,8 @@ class SKDeleteButton: SKButton {
         showFrame = CGRect(x: SKMesurement.screenWidth - size.width, y: buttonTopOffset, width: size.width, height: size.height)
         hideFrame = CGRect(x: SKMesurement.screenWidth - size.width, y: -20, width: size.width, height: size.height)
     }
-    
-    override func updateFrame() {
-    }
   
-    override func setFrameSize(size: CGSize) {
+    override func setFrameSize(_ size: CGSize) {
         let newRect = CGRect(x: SKMesurement.screenWidth - size.width, y: buttonTopOffset, width: size.width, height: size.height)
         self.frame = newRect
         showFrame = newRect