+++ SKPhotoBrowser/SKPhoto.swift
@@ -18,14 +18,14 @@ import UIKit
 }
 
 // MARK: - SKPhoto
-public class SKPhoto: NSObject, SKPhotoProtocol {
+open class SKPhoto: NSObject, SKPhotoProtocol {
     
-    public var underlyingImage: UIImage!
-    public var photoURL: String!
-    public var contentMode: UIViewContentMode = .ScaleAspectFill
-    public var shouldCachePhotoURLImage: Bool = false
-    public var caption: String!
-    public var index: Int = 0
+    open var underlyingImage: UIImage!
+    open var photoURL: String!
+    open var contentMode: UIViewContentMode = .scaleAspectFill
+    open var shouldCachePhotoURLImage: Bool = false
+    open var caption: String!
+    open var index: Int = 0
 
     override init() {
         super.init()
@@ -47,7 +47,7 @@ public class SKPhoto: NSObject, SKPhotoProtocol {
         underlyingImage = holder
     }
     
-    public func checkCache() {
+    open func checkCache() {
         guard let photoURL = photoURL else {
             return
         }
@@ -56,7 +56,7 @@ public class SKPhoto: NSObject, SKPhotoProtocol {
         }
         
         if SKCache.sharedCache.imageCache is SKRequestResponseCacheable {
-            let request = NSURLRequest(URL: NSURL(string: photoURL)!)
+            let request = URLRequest(url: URL(string: photoURL)!)
             if let img = SKCache.sharedCache.imageForRequest(request) {
                 underlyingImage = img
             }
@@ -67,50 +67,44 @@ public class SKPhoto: NSObject, SKPhotoProtocol {
         }
     }
     
-    public func loadUnderlyingImageAndNotify() {
+    open func loadUnderlyingImageAndNotify() {
+        guard photoURL != nil, let URL = URL(string: photoURL) else { return }
         
-        if underlyingImage != nil {
-            loadUnderlyingImageComplete()
-            return
-        }
-        
-        if photoURL != nil {
-            // Fetch Image
-            let session = NSURLSession(configuration: NSURLSessionConfiguration.defaultSessionConfiguration())
-            if let nsURL = NSURL(string: photoURL) {
-                var task: NSURLSessionDataTask!
-                task = session.dataTaskWithURL(nsURL, completionHandler: { [weak self](response: NSData?, data: NSURLResponse?, error: NSError?) in
-                    if let _self = self {
-                        
-                        if error != nil {
-                            dispatch_async(dispatch_get_main_queue()) {
-                                _self.loadUnderlyingImageComplete()
-                            }
-                        }
-                        
-                        if let res = response, image = UIImage(data: res) {
-                            if _self.shouldCachePhotoURLImage {
-                                if SKCache.sharedCache.imageCache is SKRequestResponseCacheable {
-                                    SKCache.sharedCache.setImageData(response!, response: data!, request: task.originalRequest!)
-                                } else {
-                                    SKCache.sharedCache.setImage(image, forKey: _self.photoURL)
-                                }
-                            }
-                            dispatch_async(dispatch_get_main_queue()) {
-                                _self.underlyingImage = image
-                                _self.loadUnderlyingImageComplete()
-                            }
-                        }
-                        session.finishTasksAndInvalidate()
+        // Fetch Image
+        let session = URLSession(configuration: URLSessionConfiguration.default)
+            var task: URLSessionTask?
+            task = session.dataTask(with: URL, completionHandler: { [weak self] (data, response, error) in
+                guard let `self` = self else { return }
+                
+                defer { session.finishTasksAndInvalidate() }
+
+                guard error == nil else {
+                    DispatchQueue.main.async {
+                        self.loadUnderlyingImageComplete()
                     }
-                })
-                task.resume()
-            }
-        }
+                    return
+                }
+
+                if let data = data, let response = response, let image = UIImage(data: data) {
+                    if self.shouldCachePhotoURLImage {
+                        if SKCache.sharedCache.imageCache is SKRequestResponseCacheable {
+                            SKCache.sharedCache.setImageData(data, response: response, request: task?.originalRequest)
+                        } else {
+                            SKCache.sharedCache.setImage(image, forKey: self.photoURL)
+                        }
+                    }
+                    DispatchQueue.main.async {
+                        self.underlyingImage = image
+                        self.loadUnderlyingImageComplete()
+                    }
+                }
+                
+            })
+            task?.resume()
     }
 
-    public func loadUnderlyingImageComplete() {
-        NSNotificationCenter.defaultCenter().postNotificationName(SKPHOTO_LOADING_DID_END_NOTIFICATION, object: self)
+    open func loadUnderlyingImageComplete() {
+        NotificationCenter.default.post(name: Notification.Name(rawValue: SKPHOTO_LOADING_DID_END_NOTIFICATION), object: self)
     }
     
 }
@@ -118,15 +112,15 @@ public class SKPhoto: NSObject, SKPhotoProtocol {
 // MARK: - Static Function
 
 extension SKPhoto {
-    public static func photoWithImage(image: UIImage) -> SKPhoto {
+    public static func photoWithImage(_ image: UIImage) -> SKPhoto {
         return SKPhoto(image: image)
     }
     
-    public static func photoWithImageURL(url: String) -> SKPhoto {
+    public static func photoWithImageURL(_ url: String) -> SKPhoto {
         return SKPhoto(url: url)
     }
     
-    public static func photoWithImageURL(url: String, holder: UIImage?) -> SKPhoto {
+    public static func photoWithImageURL(_ url: String, holder: UIImage?) -> SKPhoto {
         return SKPhoto(url: url, holder: holder)
     }
 }