+++ SKPhotoBrowser/SKPhotoBrowserOptions.swift
@@ -30,14 +30,18 @@ public struct SKPhotoBrowserOptions {
     public static var enableZoomBlackArea: Bool = true
     public static var enableSingleTapDismiss: Bool = false
     
-    public static var backgroundColor = UIColor.blackColor()
-    public static var textAndIconColor = UIColor.whiteColor()
-    public static var toolbarTextShadowColor = UIColor.darkTextColor()
-    
-    public static var toolbarFont = UIFont(name: "Helvetica", size: 16.0)
-    public static var captionFont = UIFont.systemFontOfSize(17.0)
-    
-    // FIXED: Scrolling performance slowed #145
-    // public static var imagePaddingX: CGFloat = 0
-    // public static var imagePaddingY: CGFloat = 0
+    public static var backgroundColor: UIColor = .black
+}
+
+public struct SKCaptionOptions {
+    public static var textColor: UIColor = .white
+    public static var textAlignment: NSTextAlignment = .center
+    public static var numberOfLine: Int = 3
+    public static var lineBreakMode: NSLineBreakMode = .byTruncatingTail
+    public static var font: UIFont = .systemFont(ofSize: 17.0)
+}
+
+public struct SKToolbarOptions {
+    public static var textColor: UIColor = .white
+    public static var font: UIFont = .systemFont(ofSize: 17.0)
 }